Code/Chatbot/requirements.txt
llama-cpp-python
gradio
pyngrok
pyyaml
